Reduced some smaller values to uint8_ts

+++ content/OTRadioLink/utility/OTV0P2BASE_EEPROM.h
@@ -107,13 +107,13 @@ namespace OTV0P2BASE
 //TODO-793
 // 16 Byte primary building key (secret key)
 static const uint16_t VOP2BASE_EE_START_16BYTE_PRIMARY_BUILDING_KEY = 112;
-static const uint16_t VOP2BASE_EE_LEN_16BYTE_PRIMARY_BUILDING_KEY  = 16;
+static const uint8_t VOP2BASE_EE_LEN_16BYTE_PRIMARY_BUILDING_KEY  = 16;
 
 
 // Start area of radio config mem
-static const uint16_t V0P2BASE_EE_START_RADIO = 128; // INCLUSIVE START OF RADIO CONFIG AREA.
-static const uint16_t V0P2BASE_EE_LEN_RADIO  = 128; // SIZE OF RADIO CONFIG AREA.
-static const uint16_t V0P2BASE_EE_END_RADIO  = 255;
+static const uint16_t V0P2BASE_EE_START_RADIO = 128; // INCLUSIVE START OF RADIO CONFIG AREA. // FIXME should this be a uint8_t? (technically a pointer to a 1 KiB eeprom)
+static const uint8_t V0P2BASE_EE_LEN_RADIO  = 128; // SIZE OF RADIO CONFIG AREA.
+static const uint8_t V0P2BASE_EE_END_RADIO  = 255;
 
 //// Housecode filter at central hub.
 //// Intended to fit snug up before stats area.
@@ -159,14 +159,14 @@ static const uint16_t V0P2BASE_EE_END_RADIO  = 255;
 // TODO-793
 // Node associations memory. Should fit 16 nodes within 256 bytes of EEPROM
 static const uint16_t V0P2BASE_EE_START_NODE_ASSOCIATIONS = 768; // Inclusive start of node associations
-static const uint16_t V0P2BASE_EE_NODE_ASSOCIATIONS_SET_SIZE  = 16;              // Size in entries/bytes of one Node association entry
+static const uint8_t V0P2BASE_EE_NODE_ASSOCIATIONS_SET_SIZE  = 16;              // Size in entries/bytes of one Node association entry
 
 // Node association fields, 0 upwards, contiguous.
-static const uint16_t V0P2BASE_EE_NODE_ASSOCIATIONS_8B_ID_OFFSET  = 0;  // 8 Byte node ID.
-static const uint16_t V0P2BASE_EE_NODE_ASSOCIATIONS_8B_ID_LENGTH  = 8;  // 8 Byte node ID.
-static const uint16_t V0P2BASE_EE_NODE_ASSOCIATIONS_RESERVED      = 8;  // Reserved
+static const uint8_t V0P2BASE_EE_NODE_ASSOCIATIONS_8B_ID_OFFSET  = 0;  // 8 Byte node ID.
+static const uint8_t V0P2BASE_EE_NODE_ASSOCIATIONS_8B_ID_LENGTH  = 8;  // 8 Byte node ID.
+static const uint8_t V0P2BASE_EE_NODE_ASSOCIATIONS_RESERVED      = 8;  // Reserved
 
-static const uint16_t V0P2BASE_EE_NODE_ASSOCIATIONS_MAX_SETS      = 16; // Maximum possible node associations
+static const uint8_t V0P2BASE_EE_NODE_ASSOCIATIONS_MAX_SETS      = 16; // Maximum possible node associations
 
 // Compute start of node association set (in range [0,V0P2BASE_EE_NODE_ASSOCIATIONS_SETS-1]) in EEPROM.
 // static const uint16_t V0P2BASE_EE_NODE_ASSOCIATIONS_START_ADDR
